Small edit

+++ core/types/src/array_type.cpp
@@ -1,4 +1,5 @@
 #include "array_type.h"
+#include "json_item.h"
 
 #pragma region Construcors
 
@@ -8,16 +9,16 @@ Array::Array(Array &&array) noexcept : type_array(std::move(array)) {}
 
 #pragma endregion
 
-std::ostream &operator<<(std::ostream &os, Array & array) {
-    os << "[";
-    for (int i {0}; i < array.size(); ++i) {
+std::ostream & operator << (std::ostream & os, Array & array) {
+    os << '[';
+    for (size_t i {0}; i < array.size(); ++i) {
         os << array.at(i) << ((i == array.size() - 1) ? "" : ", ");
     }
-    os << "]";
+    os << ']';
     return os;
 }
 
-std::ostream &operator<<(std::ostream &os, Array && array) {
+std::ostream & operator << (std::ostream &os, Array && array) {
     os << array;
     return os;
 }
